Update PermissionRegistrar
The package assumes that the model which is being assigned roles/permissions (e.g. User) implements the Authenticatable contract. This is not always the case. Laravel itself adds its authorization method, i.e. can to the Authorizable trait, and not Authenticatable.
This commit changes the type-hinted Authenticatable to Authorizable instead. Closes #694.