Auto-refresh role/permission models on assign/remove
This continues where #916 left off, adding the refresh to not just the Assign method, but also the Remove method on roles, and to both grant/revoke on permissions.

This means there's no longer a need to call `->fresh()` on the model if the only reason is to reload the role/permission relations. (That said, you may want to call it for other reasons.)