Historically, since v2.0.0, the "detected guard_name" that was applied during lookups was detected from the "first possible defined guard" for that model, even if the "current" guard wasn't "first in the list". This has led to some confusion when expecting that the current user's guard would actually be used.
This PR changes things to **use the current user's guard** first, assuming it's in the list of potential guards acceptable for the current User model.
After that it will fallback to the old behavior of selecting the first available matchable guard for the current User model.

THIS IS A BREAKING CHANGE, so be sure to test all authorization aspects of your application to ensure desired behavior.
You MAY also be able to remove some old complex monkey-patching that you might have done to work around this issue in prior versions.

Fixes #1682
Fixes #1608
Fixes #1384
Fixes #1515
Fixes #1516